most of crosscompilation, opencvs is a problem
