+++ jvm/src/test/scala/purecsv/safe/TryUtilSuite.scala
