+++ shared/src/main/scala/purecsv/safe/converter/defaults/string/package.scala
