+++ shared/src/main/scala/purecsv/safe/tryutil/TryUtil.scala
