+++ shared/src/main/scala/purecsv/unsafe/RecordSplitter.scala
@@ -45,37 +45,3 @@ trait RecordSplitter[R] {
     getRecords(r, RecordSplitter.defaultFieldSeparator, RecordSplitter.defaultQuoteChar, 1)
   }
 }
-
-/**
- * A [[RecordSplitter]] that uses the OpenCSV library for extracting records from a [[Reader]]
- */
-object OpenCSVSplitter extends RecordSplitter[Reader] {
-
-  override def getRecords(reader: Reader, fieldSep: Char, quoteChar: Char, firstLine: Int): Iterator[Seq[String]] = {
-    val csvReader = new au.com.bytecode.opencsv.CSVReader(reader, fieldSep, quoteChar, firstLine)
-
-    new Iterator[Seq[String]] {
-      private var nextRecord = {
-        val r = csvReader.readNext()
-        if (r == null) {
-          csvReader.close()
-        }
-        r
-      }
-
-      def hasNext(): Boolean = (nextRecord != null)
-      def next(): Seq[String] = {
-        if (nextRecord == null) {
-          throw new NoSuchElementException()
-        } else {
-          val nextRecordBuf = this.nextRecord
-          this.nextRecord = csvReader.readNext()
-          if (!this.hasNext()) {
-            csvReader.close()
-          }
-          nextRecordBuf
-        }
-      }
-    }
-  }
-}
